code/requirements.txt
torch
torchvision
numpy
Pillow
matplotlib
lpips
pandas
PyYAML
